% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastHurdleNegBinomRegression.R
\docType{package}
\name{fastHurdleNegBinomRegression}
\alias{fastHurdleNegBinomRegression}
\title{Fast Hurdle Inflated Negative Binomial Model Fitter}
\description{
A tool to find many types of a priori experimental designs
}
\references{
Kapelner, A
}
\author{
Adam Kapelner \email{kapelner@qc.cuny.edu}
}
\keyword{binomial}
\keyword{hurdle}
\keyword{negative}
